fix(lab5):extend 4MB page mapping when initialize memory
The size of kernel(including page table as well) is reaching the
boundary of 4MB, so the kernel will reboot infinitely due to
triple-fault(page fault). We setup another 'extended' page mapping
during boot, and it seems to be enough for kernel initializing itself.

BREAKING CHANGES:the initial page mapping has changed to [0, 8MB), data
and code below this is likely to be corrupted.